+++ bankAccount/src/main/kotlin/com/example/bankAccount/adapters/dto/ViewBalanceResponse.kt
@@ -1,3 +1,5 @@
 package com.example.bankAccount.adapters.dto
 
-data class ViewBalanceResponse(var balance: Double)
+import java.math.BigDecimal
+
+data class ViewBalanceResponse(var balance: BigDecimal)
