+++ bankAccount/src/main/kotlin/com/example/bankAccount/adapters/dto/WithdrawMoneyRequest.kt
@@ -1,3 +1,5 @@
 package com.example.bankAccount.adapters.dto
 
-data class WithdrawMoneyRequest(val accountId: Long, val amount: Double)
+import java.math.BigDecimal
+
+data class WithdrawMoneyRequest(val accountId: Long, val amount: BigDecimal)
