+++ bankAccount/src/main/kotlin/com/example/bankAccount/domain/ports/in/WithdrawMoneyInput.kt
@@ -1,5 +1,7 @@
 package com.example.bankAccount.domain.ports.`in`
 
+import java.math.BigDecimal
+
 interface WithdrawMoneyInput {
-    fun withdrawMoney(accountId: Long, amount: Double)
+    fun withdrawMoney(accountId: Long, amount: BigDecimal)
 }
